added new class CheckboxExamples.cs (extracted tests from ApiExamples.cs)
